+++ plugins/01_WorldOfWarcraft/localization/esES.lua
@@ -20,6 +20,7 @@ L['in_crypt'] = 'En la cripta.'
 L['shovelphlange_location'] = 'En la cueva con la entrada a Uldaman.'
 L['lower_floor'] = 'En el piso inferior.'
 L['bl_rare_note'] = 'Todos los raros de {location:Las Tierras Devastadas} pueden soltar tanto la {item:8244} como {item:10593} que puedes intercambiar por objetos de {npc:7363}.'
+L['takk_note'] = 'Literalmente salta de un lugar a otro (en sentido antihorario).'
 
 -------------------------------------------------------------------------------
 ----------------------------- ZUL'GURUB (10.0.7) ------------------------------