+++ plugins/11_TheWarWithin/localization/enUS.lua
@@ -587,8 +587,20 @@ L['impotent_potable_location'] = 'Below the dock next to the {location:Plunder P
 L['abandoned_toolbox_location'] = 'High on the roof of {location:The Scrapshop}.\n\nFollow the stairs to the top.'
 L['trick_deck_of_cards_location'] = 'On the desk next to {npc:231045} inside {location:The Incontinental Hotel}.'
 
+L['breakneck_customization'] = 'G-99 Breakneck'
+
 L['ditty_fuzeboy_note'] = 'Exchange {item:234741} for pets.'
-L['angelo_rustbin_note'] = 'Exchange {currency:3220} for pets and {object:G-99 Breakneck} customizations.'
+L['angelo_rustbin_note'] = 'Exchange {currency:3218} for a toy and a {object:G-99 Breakneck} customization. Exchange {item:233557} for a chance to receive pet and mount.\n\nExchange {currency:3220} for pets.\n\n{note:Collecting {currency:3218} can done at any time but collecting {item:229810} requires {faction:2653} Renown 14.}'
+L['blair_bass_note'] = 'Exchange {item:227673} for toys.\n\n{item:227673} can be fished up from {object:Steamwheedle Runoff} pools.'
+L['greexit_coarsebub_note'] = 'Exchange gold for goblin cartel transmogs.\n\n{note:Only items associated with your current cartel are available to purchase.}'
+
+L['quartermaster'] = '%s Quartermaster'
+L['quartermaster_note'] = 'Exchange {currency:2815} for a variety of collectibles.'
+
+L['rocco_razzboom_location'] = 'Bottom floor of {location:The Scrapshop}.'
+L['lab_assistant_laszly_location'] = 'Deep inside {location:Steamwheedle Laboratories}.'
+L['boatswain_hardee_location'] = 'First floor of {location:Port Authority}.'
+L['shrez_the_scrapper_location'] = 'First floor of {location:Venture Plaza}.'
 
 L['options_icons_nine_tenths'] = '{achievement:40948}'
 L['options_icons_nine_tenths_desc'] = 'Display all {object:Muff\'s Auto-Locker} locations for {achievement:40948}.'
@@ -615,6 +627,13 @@ L['options_icons_youre_my_friend_now'] = '{achievement:41708}'
 L['options_icons_youre_my_friend_now_desc'] = 'Display {npc:239248} locations for {achievement:41708}.'
 L['rat_in_building_rail'] = 'Inside the building on the rail.'
 L['rat_2nd_floor_trash_can'] = 'Inside the building on the second floor on the trash can.'
+
+L['sewer_cheese_label'] = 'Sewer Cheese'
+L['undermine_undershirt_note'] = '1. Collect the {object:Sewer Cheese} from within the sewer below {location:The Vatworks} to receive the {spell:1221472} buff.\n\n2. Interact with the nearby {npc:238661} to trade your cheese for a {item:237129}.\n\n3. Bring the {item:237129} to {npc:237412}, upstairs in the {location:Port Authority} building at {location:Blackwater Marina}, and exchange it for {item:237130}.'
+
+L['undermine_clam_chowder_label'] = 'The Undermine Guide to Authentic Cooking'
+L['undermine_clam_chowder_location'] = 'On top of a shelf behind {npc:226728} in the alley in {location:The Vatworks}.'
+L['undermine_clam_chowder_note'] = '1. To collect this recipe you must first know {item:16767} which can be purchased from two {object:Classic} vendors for %s.\n\n{npc:54232} at {location:Booty Bay} in {location:The Cape of Stranglethorn}\n\n{npc:48060} at {location:Fuselight} in {location:Badlands}\n\n2. Once you know the old recipe, open {object:The Undermine Guide to Authentic Cooking} and read to the last page. Jot down the recipe to recieve {item:235800}.'
 
 L['breakneck_race_note'] = 'The following {object:G-99 Breakneck} upgrades are recommended to make earning |cFFFFD700Gold|r a little easier:\n\n%s\n\n%s\n\n%s'
 L['upgrade_the_pozzik_standard'] = '{item:232982} (Engine): Earn |cFFCD7F32Bronze|r in all normal and reverse {object:Breakneck} races.'