+++ plugins/11_TheWarWithin/localization/zhTW.lua
@@ -587,8 +587,20 @@ L['impotent_potable_location'] = nil
 L['abandoned_toolbox_location'] = nil
 L['trick_deck_of_cards_location'] = nil
 
+L['breakneck_customization'] = nil
+
 L['ditty_fuzeboy_note'] = nil
 L['angelo_rustbin_note'] = nil
+L['blair_bass_note'] = nil
+L['greexit_coarsebub_note'] = nil
+
+L['quartermaster'] = nil
+L['quartermaster_note'] = nil
+
+L['rocco_razzboom_location'] = nil
+L['lab_assistant_laszly_location'] = nil
+L['boatswain_hardee_location'] = nil
+L['shrez_the_scrapper_location'] = nil
 
 L['options_icons_nine_tenths'] = '{achievement:40948}'
 L['options_icons_nine_tenths_desc'] = nil
@@ -615,6 +627,13 @@ L['options_icons_youre_my_friend_now'] = '{achievement:41708}'
 L['options_icons_youre_my_friend_now_desc'] = nil
 L['rat_in_building_rail'] = nil
 L['rat_2nd_floor_trash_can'] = nil
+
+L['sewer_cheese_label'] = nil
+L['undermine_undershirt_note'] = nil
+
+L['undermine_clam_chowder_label'] = nil
+L['undermine_clam_chowder_location'] = nil
+L['undermine_clam_chowder_note'] = nil
 
 L['breakneck_race_note'] = nil
 L['upgrade_the_pozzik_standard'] = nil