+++ plugins/11_TheWarWithin/zones/undermine.lua
@@ -18,6 +18,8 @@ local Item = ns.reward.Item
 local Mount = ns.reward.Mount
 local Pet = ns.reward.Pet
 local Reputation = ns.reward.Reputation
+local Section = ns.reward.Section
+local Spacer = ns.reward.Spacer
 local Toy = ns.reward.Toy
 local Transmog = ns.reward.Transmog
 
@@ -40,9 +42,11 @@ map.nodes[68508078] = Rare({
     id = 230931,
     quest = 84917,
     rewards = {
-        Achievement({id = 41216, criteria = 71591}), --
-        Item({item = 234419}), -- Steamboil Mounting System
-        Reputation({id = 2653, gain = 20})
+        Achievement({id = 41216, criteria = 71591}),
+        Reputation({id = 2653, gain = 20}),
+        Transmog({item = 235301, slot = L['cloak']}), -- Drape of the Dazzling Feather
+        Transmog({item = 235305, slot = L['cloth']}), -- Golfer's Truestrike Gloves
+        Transmog({item = 235321, slot = L['mail']}) -- Feather-Spike Girdle
     }
 }) -- Scrapbeak
 
@@ -53,7 +57,9 @@ map.nodes[25243675] = Rare({
     quest = 84918,
     rewards = {
         Achievement({id = 41216, criteria = 71592}),
-        Reputation({id = 2653, gain = 20})
+        Reputation({id = 2653, gain = 20}),
+        Transmog({item = 235308, slot = L['cloth']}), -- Filthtread Boots
+        Transmog({item = 235326, slot = L['leather']}) -- Ratspit's Heirloom Wristwraps
     }
 }) -- Ratspit
 
@@ -61,9 +67,15 @@ map.nodes[37684448] = Rare({
     id = 230940,
     quest = 84919,
     rewards = {
-        Achievement({id = 41216, criteria = 71593}), --
-        Item({item = 234420}), -- Handcrank Mounting System
-        Reputation({id = 2653, gain = 20})
+        Achievement({id = 41216, criteria = 71593}),
+        Reputation({id = 2653, gain = 20}),
+        Transmog({item = 235328, slot = L['leather']}), -- Boots of the Silver Tongue
+        Transmog({item = 235355, slot = L['staff']}), -- Gossi-blin's Baton
+        Spacer(), Section(L['shared_drops']),
+        Transmog({item = 235310, slot = L['cloth']}), -- Flashy Patchwork Trousers
+        Transmog({item = 235320, slot = L['mail']}), -- S.1.Z.Z.L.E.S.T.E.P Boots
+        Transmog({item = 235327, slot = L['leather']}), -- Mend-and-Match Shoulderpads
+        Transmog({item = 235347, slot = L['dagger']}) -- 100% Sharp Glimmerblade
     }
 }) -- Tally Doublespeak
 
@@ -73,7 +85,14 @@ map.nodes[36877815] = Rare({
     sublabel = L['and_slimesby'],
     rewards = {
         Achievement({id = 41216, criteria = {71595, 71594}}),
-        Reputation({id = 2653, gain = 20})
+        Reputation({id = 2653, gain = 20}),
+        Transmog({item = 235306, slot = L['cloth']}), -- Ooze-fused Mantle
+        Transmog({item = 235329, slot = L['leather']}), -- Cowl of Acidic Mire
+        Spacer(), Section(L['shared_drops']),
+        Transmog({item = 235310, slot = L['cloth']}), -- Flashy Patchwork Trousers
+        Transmog({item = 235320, slot = L['mail']}), -- S.1.Z.Z.L.E.S.T.E.P Boots
+        Transmog({item = 235327, slot = L['leather']}), -- Mend-and-Match Shoulderpads
+        Transmog({item = 235347, slot = L['dagger']}) -- 100% Sharp Glimmerblade
     }
 }) -- V.V. Goosworth and Slimesby
 
@@ -81,9 +100,15 @@ map.nodes[54015023] = Rare({
     id = 230951,
     quest = 84921,
     rewards = {
-        Achievement({id = 41216, criteria = 71596}), --
-        Item({item = 234420}), -- Handcrank Mounting System
-        Reputation({id = 2653, gain = 20})
+        Achievement({id = 41216, criteria = 71596}),
+        Reputation({id = 2653, gain = 20}),
+        Transmog({item = 235317, slot = L['plate']}), -- Chestplate of the Ultimatum
+        Transmog({item = 235353, slot = L['1h_axe']}), -- Debtsmasher Axe
+        Spacer(), Section(L['shared_drops']),
+        Transmog({item = 235310, slot = L['cloth']}), -- Flashy Patchwork Trousers
+        Transmog({item = 235320, slot = L['mail']}), -- S.1.Z.Z.L.E.S.T.E.P Boots
+        Transmog({item = 235327, slot = L['leather']}), -- Mend-and-Match Shoulderpads
+        Transmog({item = 235347, slot = L['dagger']}) -- 100% Sharp Glimmerblade
     }
 }) -- Thwack
 
@@ -91,9 +116,14 @@ map.nodes[41922563] = Rare({
     id = 230979,
     quest = 84922,
     rewards = {
-        Achievement({id = 41216, criteria = 71597}), --
-        Item({item = 234381}), -- Handcrank Fuel Tank
-        Reputation({id = 2653, gain = 20})
+        Achievement({id = 41216, criteria = 71597}),
+        Reputation({id = 2653, gain = 20}),
+        Transmog({item = 235356, slot = L['offhand']}), -- Sapper's Spark Reactor
+        Spacer(), Section(L['shared_drops']),
+        Transmog({item = 235300, slot = L['cloak']}), -- Cloak of Mecha Shards
+        Transmog({item = 235315, slot = L['plate']}), -- Rocketstep Boots
+        Transmog({item = 235322, slot = L['mail']}), -- Junkyard Clawguards
+        Transmog({item = 235351, slot = L['1h_sword']}) -- Hypersteel CX4 Greatsword
     }
 }) -- S.A.L.
 
@@ -102,8 +132,10 @@ map.nodes[46915565] = Rare({
     quest = 84926,
     rewards = {
         Achievement({id = 41216, criteria = 71598}),
+        Reputation({id = 2653, gain = 20}),
+        Transmog({item = 235318, slot = L['plate']}), -- Ironfang Plate Legguards
         Transmog({item = 235324, slot = L['mail']}), -- Scavenger's Lost Bind
-        Reputation({id = 2653, gain = 20})
+        Transmog({item = 235325, slot = L['leather']}) -- Rusthide Gloves
     }
 }) -- Nitro
 
@@ -112,7 +144,14 @@ map.nodes[42227735] = Rare({
     quest = 84927,
     rewards = {
         Achievement({id = 41216, criteria = 71599}),
-        Reputation({id = 2653, gain = 20})
+        Reputation({id = 2653, gain = 20}),
+        Transmog({item = 235304, slot = L['cloth']}), -- Gutter Rat Mask
+        Transmog({item = 235348, slot = L['dagger']}), -- Back Alley Shank
+        Spacer(), Section(L['shared_drops']),
+        Transmog({item = 235310, slot = L['cloth']}), -- Flashy Patchwork Trousers
+        Transmog({item = 235320, slot = L['mail']}), -- S.1.Z.Z.L.E.S.T.E.P Boots
+        Transmog({item = 235327, slot = L['leather']}), -- Mend-and-Match Shoulderpads
+        Transmog({item = 235347, slot = L['dagger']}) -- 100% Sharp Glimmerblade
     }
 }) -- Candy Stickemup
 
@@ -121,7 +160,9 @@ map.nodes[67333353] = Rare({
     quest = 84928,
     rewards = {
         Achievement({id = 41216, criteria = 71600}),
-        Reputation({id = 2653, gain = 20})
+        Reputation({id = 2653, gain = 20}),
+        Transmog({item = 235319, slot = L['mail']}), -- Tidebomb Chestpiece
+        Transmog({item = 235323, slot = L['mail']}) -- Blastshell Bracers
     }
 }) -- Grimewick
 
@@ -130,7 +171,14 @@ map.nodes[41354357] = Rare({
     quest = 85004,
     rewards = {
         Achievement({id = 41216, criteria = 71601}),
-        Reputation({id = 2653, gain = 20})
+        Reputation({id = 2653, gain = 20}),
+        Transmog({item = 235307, slot = L['cloth']}), -- Smoketrail Belt
+        Transmog({item = 235314, slot = L['plate']}), -- Knightrider's Steelfists
+        Spacer(), Section(L['shared_drops']),
+        Transmog({item = 235310, slot = L['cloth']}), -- Flashy Patchwork Trousers
+        Transmog({item = 235320, slot = L['mail']}), -- S.1.Z.Z.L.E.S.T.E.P Boots
+        Transmog({item = 235327, slot = L['leather']}), -- Mend-and-Match Shoulderpads
+        Transmog({item = 235347, slot = L['dagger']}) -- 100% Sharp Glimmerblade
     }
 }) -- Swigs Farsight
 
@@ -139,7 +187,10 @@ map.nodes[26516830] = Rare({
     quest = 84877,
     rewards = {
         Achievement({id = 41216, criteria = 71602}),
-        Reputation({id = 2653, gain = 50})
+        Reputation({id = 2653, gain = 50}),
+        Transmog({item = 235309, slot = L['cloth']}), -- Gloomshroud Robe
+        Transmog({item = 235350, slot = L['1h_mace']}), -- Void-forged Cudgel
+        Transmog({item = 235352, slot = L['wand']}) -- Netherflare Wand
     }
 }) -- Ephemeral Agent Lathyd
 
@@ -147,9 +198,15 @@ map.nodes[63354975] = Rare({
     id = 230793,
     quest = 84884,
     rewards = {
-        Achievement({id = 41216, criteria = 71603}), --
-        Item({item = 234419}), -- Steamboil Mounting System
-        Reputation({id = 2653, gain = 50})
+        Achievement({id = 41216, criteria = 71603}),
+        Reputation({id = 2653, gain = 50}),
+        Transmog({item = 235313, slot = L['plate']}), -- Shockproof Helm
+        Transmog({item = 235354, slot = L['polearm']}), -- Scrapblaster Lance
+        Spacer(), Section(L['shared_drops']),
+        Transmog({item = 235300, slot = L['cloak']}), -- Cloak of Mecha Shards
+        Transmog({item = 235315, slot = L['plate']}), -- Rocketstep Boots
+        Transmog({item = 235322, slot = L['mail']}), -- Junkyard Clawguards
+        Transmog({item = 235351, slot = L['1h_sword']}) -- Hypersteel CX4 Greatsword
     },
     pois = {Path({63574989, 64085016, 64005291})}
 }) -- The Junk-Wall
@@ -159,7 +216,13 @@ map.nodes[52354107] = Rare({
     quest = 84895,
     rewards = {
         Achievement({id = 41216, criteria = 71604}),
-        Reputation({id = 2653, gain = 50})
+        Reputation({id = 2653, gain = 20}),
+        Transmog({item = 235349, slot = L['1h_mace']}), -- Shadowfume Club
+        Spacer(), Section(L['shared_drops']),
+        Transmog({item = 235310, slot = L['cloth']}), -- Flashy Patchwork Trousers
+        Transmog({item = 235320, slot = L['mail']}), -- S.1.Z.Z.L.E.S.T.E.P Boots
+        Transmog({item = 235327, slot = L['leather']}), -- Mend-and-Match Shoulderpads
+        Transmog({item = 235347, slot = L['dagger']}) -- 100% Sharp Glimmerblade
     }
 }) -- Slugger the Smart
 
@@ -168,14 +231,30 @@ map.nodes[58488643] = Rare({
     quest = 84907,
     rewards = {
         Achievement({id = 41216, criteria = 71605}),
-        Reputation({id = 2653, gain = 50})
+        Reputation({id = 2653, gain = 50}),
+        Transmog({item = 235311, slot = L['plate']}), -- Rocket-Powered Shoulderguards
+        Transmog({item = 235357, slot = L['shield']}), -- Bulletscar Barricade
+        Spacer(), Section(L['shared_drops']),
+        Transmog({item = 235300, slot = L['cloak']}), -- Cloak of Mecha Shards
+        Transmog({item = 235315, slot = L['plate']}), -- Rocketstep Boots
+        Transmog({item = 235322, slot = L['mail']}), -- Junkyard Clawguards
+        Transmog({item = 235351, slot = L['1h_sword']}) -- Hypersteel CX4 Greatsword
     }
 }) -- Chief Foreman Gutso
 
 map.nodes[60580989] = Rare({
     id = 230840,
     quest = 84911,
-    rewards = {Achievement({id = 41216, criteria = 71606})},
+    rewards = {
+        Achievement({id = 41216, criteria = 71606}),
+        Transmog({item = 235312, slot = L['plate']}), -- Snooty's Aviator Bindings
+        Transmog({item = 235316, slot = L['plate']}), -- Whirly-Giggle Windwhir Wrap
+        Spacer(), Section(L['shared_drops']),
+        Transmog({item = 235300, slot = L['cloak']}), -- Cloak of Mecha Shards
+        Transmog({item = 235315, slot = L['plate']}), -- Rocketstep Boots
+        Transmog({item = 235322, slot = L['mail']}), -- Junkyard Clawguards
+        Transmog({item = 235351, slot = L['1h_sword']}) -- Hypersteel CX4 Greatsword
+    },
     vignette = 6597
 }) -- Flyboy Snooty
 
@@ -188,7 +267,13 @@ map.nodes[40002232] = Rare({
     },
     rewards = {
         Achievement({id = 41216, criteria = 71608}),
-        Transmog({item = 235318, type = L['plate']}) -- Ironfang Plate Legguards
+        Transmog({item = 235835, type = L['cloak']}), -- Braided Wire Wrap
+        Transmog({item = 235836, type = L['2h_sword']}), -- Gas-Powered Chainblade
+        Spacer(), Section(L['shared_drops']),
+        Transmog({item = 235300, slot = L['cloak']}), -- Cloak of Mecha Shards
+        Transmog({item = 235315, slot = L['plate']}), -- Rocketstep Boots
+        Transmog({item = 235322, slot = L['mail']}), -- Junkyard Clawguards
+        Transmog({item = 235351, slot = L['1h_sword']}) -- Hypersteel CX4 Greatsword
     },
     sublabel = format(L['requires_ally'], 2673) -- Bilgewater Cartel
 }) -- M.A.G.N.O.
@@ -200,7 +285,16 @@ map.nodes[32027652] = Rare({
     requires = {
         ns.requirement.Reputation(2653, 6, true) -- The Cartels of Undermine
     },
-    rewards = {Achievement({id = 41216, criteria = 71609})},
+    rewards = {
+        Achievement({id = 41216, criteria = 71609}),
+        Transmog({item = 235823, type = L['leather']}), -- Scrap-Plated Pants
+        Transmog({item = 235824, type = L['gun']}), -- Flame Sputterer
+        Spacer(), Section(L['shared_drops']),
+        Transmog({item = 235310, slot = L['cloth']}), -- Flashy Patchwork Trousers
+        Transmog({item = 235320, slot = L['mail']}), -- S.1.Z.Z.L.E.S.T.E.P Boots
+        Transmog({item = 235327, slot = L['leather']}), -- Mend-and-Match Shoulderpads
+        Transmog({item = 235347, slot = L['dagger']}) -- 100% Sharp Glimmerblade
+    },
     sublabel = format(L['requires_ally'], 2677) -- Steamwheedle Cartel
 }) -- Giovante
 
@@ -213,7 +307,9 @@ map.nodes[57207860] = Rare({
     },
     rewards = {
         Achievement({id = 41216, criteria = 71610}),
-        Transmog({item = 235829, type = L['mail']}) -- Welded Scrap Hood
+        Transmog({item = 235829, type = L['mail']}), -- Welded Scrap Hood
+        Transmog({item = 235830, type = L['mail']}), -- Unstable Missilecaps
+        Transmog({item = 235831, type = L['1h_sword']}) -- Battery-Powered Longshank
     },
     sublabel = format(L['requires_ally'], 2671) -- Venture Company
 }) -- Scrapchewer
@@ -227,7 +323,9 @@ map.nodes[64162556] = Rare({
     },
     rewards = {
         Achievement({id = 41216, criteria = 71611}),
-        Transmog({item = 235827, type = L['leather']}) -- Statically Charged Vest
+        Transmog({item = 235826, type = L['cloth']}), -- Electric Wristrags
+        Transmog({item = 235827, type = L['leather']}), -- Statically Charged Vest
+        Transmog({item = 235828, type = L['staff']}) -- Electrocution Warning
     },
     sublabel = format(L['requires_ally'], 2675) -- Blackwater Cartel
 }) -- Volstrike the Charged
@@ -235,13 +333,16 @@ map.nodes[64162556] = Rare({
 map.nodes[40209190] = Rare({
     id = 231310,
     note = format(L['complete_event'], 231329), -- De-Pollution Station X1119
-    quest = nil,
+    quest = 90492,
     requires = {
         ns.requirement.Reputation(2653, 6, true), -- The Cartels of Undermine
         ns.requirement.Item(229823) -- Canister of Darkfuse Solution
     },
     rewards = {
         Achievement({id = 41216, criteria = 71612}),
+        Transmog({item = 235832, type = L['cloak']}), -- Oil-Splattered Cloak
+        Transmog({item = 235833, type = L['polearm']}), -- Serrated Slickgrip
+        Transmog({item = 235834, type = L['warglaive']}), -- Rocketgrip Turboslicer
         Mount({item = 229955, id = 2293}) -- Darkfuse Spy-Eye
     }
 }) -- Darkfuse Precipitant
@@ -301,8 +402,7 @@ map.nodes[59351912] = Treasure({
     pois = {POI({58751798, color = 'Blue'})},
     quest = {85860, 85858},
     rewards = {
-        Achievement({id = 41217, criteria = 71626}),
-        Transmog({item = 235279, slot = L['legs']}) -- Scorched Shorts
+        Achievement({id = 41217, criteria = 71626}), Transmog({item = 235279}) -- Scorched Shorts
     }
 }) -- Lonely Tub
 
@@ -375,13 +475,16 @@ map.nodes[42308231] = Treasure({
 ----------------------------------- VENDORS -----------------------------------
 -------------------------------------------------------------------------------
 
+local BreakneckCustomization = Class('BreakneckCustomization', Item,
+    {type = L['breakneck_customization']})
+
 map.nodes[35384142] = Vendor({
     id = 236411,
     note = L['ditty_fuzeboy_note'],
     rewards = {
         Pet({item = 232850, id = 4649, count = 5}), -- Blackwater Kegmover
         Pet({item = 232846, id = 4648, count = 5}), -- Steamwheedle Flunkie
-        Pet({item = 232849, id = 4650, count = 5}), -- venture Companyman
+        Pet({item = 232849, id = 4650, count = 5}), -- Venture Companyman
         Pet({item = 232840, id = 4661, count = 5}), -- Mechagopher
         Pet({item = 232841, id = 4644, count = 8}), -- Professor Punch
         Pet({item = 232842, id = 4638, count = 10}) -- Crimson Mechasaur
@@ -392,11 +495,147 @@ map.nodes[25743813] = Vendor({
     id = 234776,
     note = L['angelo_rustbin_note'],
     rewards = {
+        Section('{currency:3218}'), -- Empty Kaja'Cola Can
+        Toy({item = 237382, count = 999}), -- Undermine Supply Crate
+        BreakneckCustomization({item = 232986, count = 2000, quest = 85781}), -- GE86 Advance
+        Spacer(), Section('{item:233557}'), -- Sifted Pile of Scrap
+        Pet({item = 232838, id = 4639}), -- Viridian Mechasaur
+        Mount({item = 229949, id = 2288}), -- Personalized Goblin S.C.R.A.P.per
+        Spacer(), Section('{currency:3220}'), -- Vintage Kaja'Cola Can
         Pet({item = 232858, id = 4655, count = 1}), -- Cruncher
-        Pet({item = 232859, id = 4653, count = 3}), -- Lab Rat
-        Item({item = 232986, count = 1}) -- GE86 Advance
+        Pet({item = 232859, id = 4653, count = 3}) -- Lab Rat
     }
 }) -- Angelo Rustbin
+
+map.nodes[34097126] = Vendor({
+    id = 226994,
+    note = L['blair_bass_note'],
+    rewards = {
+        Toy({item = 237346, count = 5}), -- Artisan Beverage Goblet Bobber
+        Toy({item = 237347, count = 5}), -- Organically-Sourced Wellington Bobber
+        Toy({item = 237345, count = 5}) -- Limited Edition Rocket Bobber
+    }
+}) -- Blair Bass
+
+map.nodes[24606320] = Vendor({
+    id = 236849,
+    note = L['greexit_coarsebub_note'],
+    rewards = {
+        Section('{faction:2673}'), -- Bilgewater
+        Transmog({item = 231742, count = 190}), -- Bilgewater Bruisers Helm
+        Transmog({item = 231748, count = 95}), -- Bilgewater Bruisers Spaulders
+        Transmog({item = 231736, count = 475}), -- Bilgewater Bruisers Tabard
+        Spacer(), --
+        Section('{faction:2677}'), -- Steamwheedle
+        Transmog({item = 231740, count = 190}), -- Steamwheedle Bruisers Helm
+        Transmog({item = 231747, count = 95}), -- Steamwheedle Bruisers Spaulders
+        Transmog({item = 231735, count = 475}), -- Steamwheedle Bruisers Tabard
+        Spacer(), --
+        Section('{faction:2675}'), -- Blackwater
+        Transmog({item = 231741, count = 190}), -- Blackwater Bruisers Helm
+        Transmog({item = 231746, count = 95}), -- Blackwater Bruisers Spaulders
+        Transmog({item = 231734, count = 475}), -- Blackwater Bruisers Tabard
+        Spacer(), --
+        Section('{faction:2671}'), -- Venture Company
+        Transmog({item = 231744, count = 190}), -- Venture Co. Bruisers Helm
+        Transmog({item = 231750, count = 95}), -- Venture Co. Bruisers Spaulders
+        Transmog({item = 231738, count = 475}) -- Venture Co. Bruisers Tabard
+    }
+}) -- Greexit Coarsebub
+
+------------------------ GOBLIN CARTEL QUARTERMASTERS -------------------------
+
+map.nodes[39152219] = Vendor({
+    id = 231406,
+    sublabel = format(L['quartermaster'], '{faction:2673}'),
+    location = L['rocco_razzboom_location'],
+    note = L['quartermaster_note'],
+    rewards = {
+        Section(L['rep_honored']),
+        BreakneckCustomization({item = 236672, quest = 85785}), -- The Ol' Low-and-Slow
+        Toy({item = 235670}), -- Bilgewater Cartel Banner
+        Spacer(), Section(L['rep_revered']), --
+        Toy({item = 235807}), -- Storefront-in-a-Box
+        Pet({item = 232845, id = 4645}), -- Bilgewater Junkhauler
+        Spacer(), Section(L['rep_exalted']),
+        BreakneckCustomization({item = 235388, quest = 86773}), -- Paint: Redlining Red
+        Transmog({item = 231526}), -- Bilgewater Undermine Tabard
+        Mount({item = 229935, id = 2272}) -- Crimson Armored Growler
+    }
+}) -- Rocco Razzboom <Bilgewater Quartermaster>
+
+map.nodes[27127256] = Vendor({
+    id = 231408,
+    sublabel = format(L['quartermaster'], '{faction:2677}'),
+    location = L['lab_assistant_laszly_location'],
+    note = L['quartermaster_note'],
+    rewards = {
+        Section(L['rep_honored']),
+        BreakneckCustomization({item = 236670, quest = 85787}), -- Maniacle Melodies
+        Toy({item = 235669}), -- Steamwheedle Cartel Banner
+        Spacer(), Section(L['rep_revered']), --
+        Toy({item = 226373}), -- Everlasting Noggenfogger Elixer
+        Pet({item = 232853, id = 4632}), -- Eepy
+        Spacer(), Section(L['rep_exalted']),
+        BreakneckCustomization({item = 235389, quest = 86772}), -- Paint: Goblin Green
+        Transmog({item = 231527}), -- Steamwheedle Undermine Tabard
+        Mount({item = 229956, id = 2294}) -- Mean Green Flying Machine
+    }
+}) -- Lab Assistant Laszly <Steamwheedle Quartermaster>
+
+map.nodes[63421673] = Vendor({
+    id = 231405,
+    sublabel = format(L['quartermaster'], '{faction:2675}'),
+    location = L['boatswain_hardee_location'],
+    note = L['quartermaster_note'],
+    rewards = {
+        Section(L['rep_honored']),
+        BreakneckCustomization({item = 236671, quest = 85786}), -- The Buzzer
+        Toy({item = 235671}), -- Blackwater Cartel Banner
+        Spacer(), Section(L['rep_revered']), --
+        Toy({item = 235801}), -- Personal Fishing Barge
+        Pet({item = 232839, id = 4637}), -- Wavebreaker Mechasaur
+        Spacer(), Section(L['rep_exalted']),
+        BreakneckCustomization({item = 235390, quest = 86771}), -- Paint: Body Roll Blue
+        Transmog({item = 231528}), -- Blackwater Undermine Tabard
+        Mount({item = 229948, id = 2286}) -- Blackwater Shredder Deluxe Mk 2
+    }
+}) -- Boatswain Hardee <Blackwater Quartermaster>
+
+map.nodes[53297271] = Vendor({
+    id = 231407,
+    sublabel = format(L['quartermaster'], '{faction:2671}'),
+    location = L['shrez_the_scrapper_location'],
+    note = L['quartermaster_note'],
+    rewards = {
+        Section(L['rep_honored']),
+        BreakneckCustomization({item = 236669, quest = 85788}), -- The Whole Brass Band
+        Toy({item = 235672}), -- Venture Co. Banner
+        Spacer(), Section(L['rep_revered']), --
+        Toy({item = 235799}), -- Throwin' Sawblade
+        Pet({item = 232851, id = 4641}), -- Rocketfist
+        Spacer(), Section(L['rep_exalted']),
+        BreakneckCustomization({item = 235391, quest = 86774}), -- Paint: Yellow Cake Yellow
+        Transmog({item = 231542}), -- Venture Co. Undermine Tabard
+        Mount({item = 229946, id = 2284}) -- Ocher Delivery Rocket
+    }
+}) -- Shredz the Scapper <Venture Quartermaster>
+
+map.nodes[30723890] = Vendor({
+    id = 231396,
+    sublabel = format(L['quartermaster'], '{faction:2669}'),
+    location = L['in_sewer'],
+    note = L['quartermaster_note'],
+    pois = {Entrance({29754112})},
+    rewards = {
+        Section(L['rep_honored']), --
+        Toy({item = 234950}), -- Atomic Regoblinator
+        Spacer(), Section(L['rep_revered']), --
+        Mount({item = 229950, id = 2287}), -- Darkfuse Demolisher
+        Spacer(), Section(L['rep_exalted']), --
+        Transmog({item = 231550}) -- Darkfuse Lowdown Coat
+    }
+}) -- Sitch Lowdown <Darkfuse Solutions Quartermaster>
 
 -------------------------------------------------------------------------------
 ----------------------------- S.C.R.A.P. REWARDS ------------------------------
@@ -410,6 +649,13 @@ local function GetScrapRewards(criteriaID)
             criteria = {id = 1, qty = true, suffix = L['scrap_heap_suffix']}
         }), -- Absolutely Zero Littering
         Achievement({id = 41594, oneline = true}),
+        Transmog({item = 236161, type = L['gun']}), -- Broiler Supreme 300
+        Transmog({item = 236178, type = L['bow']}), -- Gammy's Hand-Me-Down Bow
+        Transmog({item = 236181, type = L['1h_sword']}), -- Center-Stage Remover
+        Transmog({item = 236191, type = L['1h_mace']}), -- Mechanic's Best Motivator
+        Spacer(), Section('{npc:234621}'), -- Gallagio Garbage <Luxury Trash>
+        Transmog({item = 235821, type = L['mail']}), -- Pressed-Gold Pantaloons
+        Transmog({item = 235822, type = L['cloak']}), -- Coin-Woven Shawl
         Mount({item = 229953, id = 2291}) -- Salvaged Goblin Gazillionaire's Flying Machine
     }
 end
@@ -591,3 +837,48 @@ map.nodes[56923313] = DiscardedCan()
 
 map.nodes[43251438] = WorldsoulMemory({areaPoiID = 8200}) -- Early Cartel Wars
 map.nodes[58196864] = WorldsoulMemory({areaPoiID = 8201}) -- Kaja'mite Contact
+
+-------------------------------------------------------------------------------
+-------------------------------- MISCELLANEOUS --------------------------------
+-------------------------------------------------------------------------------
+
+---------------------------- UNDERMINE UNDERSHIRT -----------------------------
+
+map.nodes[33085816] = Collectible({
+    label = L['sewer_cheese_label'],
+    icon = 'peg_bl',
+    scale = 2.0,
+    note = L['undermine_undershirt_note'],
+    rlabel = ns.status.Gray('#1'),
+    location = L['in_sewer'],
+    pois = {Entrance({33815756})},
+    rewards = {Transmog({item = 237130})} -- Undermine Undershirt
+}) -- Sewer Cheese
+
+map.nodes[63231691] = Collectible({
+    label = '{npc:237412}',
+    icon = 'peg_bl',
+    scale = 2.0,
+    note = L['undermine_undershirt_note'],
+    rlabel = ns.status.Gray('#2'),
+    requires = ns.requirement.Item(237129), -- Tarnished Undermine Real
+    rewards = {Transmog({item = 237130})} -- Undermine Undershirt
+}) -- Pix Xizzix
+
+------------------ RECIPE: AUTHENTIC UNDERMINE CLAM CHOWDER -------------------
+
+map.nodes[38058868] = Collectible({
+    label = L['undermine_clam_chowder_label'],
+    icon = 'peg_bl',
+    scale = 2.0,
+    location = L['undermine_clam_chowder_location'],
+    note = format(L['undermine_clam_chowder_note'],
+        C_CurrencyInfo.GetCoinTextureString(3000)),
+    requires = {
+        ns.requirement.Profession(185, 2548, 225), -- Classic Cooking 225
+        ns.requirement.Profession(185, 2873, 35) -- The War Within Cooking 35
+    },
+    rewards = {
+        ns.reward.Recipe({item = 235800, profession = 185}) -- Recipe:
+    }
+}) -- Authentic Undermine Clam Chowder